infra/pyproject.toml
[project]
name = "infra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyinfra>=3.5.1",
    "python-dotenv>=1.1.1",
]
